+++ v2/slogger/rolling_file_appender/state.go
@@ -1,0 +1,5 @@
+package rolling_file_appender
+
+type state struct {
+	path string
+}
